Add client-hostname option and revert hostname -f
Adding the --client-hostname option to keycloak-httpd-client-install
commands run in the scripts to force the httpd configs to use the same
hostname used elsewhere.

Also reverting the full hostname from previous commit for hostname -f.
This is because in some environments the hostname is too long for
openssl to support (64 char limit).

Signed-off-by: Scott Poore <spoore@redhat.com>