Fix test environment issues with centos and TF
semanage was not installed on centos 10.  Added to main dnf.

In centos 10, we also saw SELinux AVCs due to the web server directory
contexts being incorrect.  Added a couple of restorecon's to resove
this.

In testing-farm environments, the hostname cannot always be resolved
properly for the Keycloak container to function.  Adding an entry to
/etc/hosts to resolve this.

In some other test environments, the hostname is set differently.
Adding another workaround to setup.sh to handle setting hostname from
reverse lookup of IP address in some cases.